+++ CHANGES.md
@@ -7,6 +7,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+Version 0.4.0 (July 31st, 2021)
+-------------------------------
+Add:
+  - `utils.get_windows_shebang` function.
+
+Changed:
+  - Support change windows shebang.
+  - Support create wheel without venv.
+
 Version 0.3.0 (July 29th, 2021)
 -------------------------------
 Add: